Trigger jobs when needed
This adds helper methods to call Special:RunJobs when needed,
to process any jobs queued by previous actions. This can be
necessary to allow tests to observe the effects of the actions
they have taken.

To avoid jobs piling up, jobs are automatically triggered when the
edit() method is called on an API client connection. This means that
all effects of the edit are visible once the await returns (barring
replication lag, see 230211).

Bug: T231822
Depends-On: Ibb38490afca71efeb67300b9665951c429c19a3c
Change-Id: I0c90eeb51c234ebc58acd18760b9cee2deaf2b6b