Make ci.sh script actually strict!
A good practice in bash script is to configure it to exit immediately
upon error (such as trying to cd to a non existing directory) or bails
out on undefined variable. That is done with bash options 'e' and 'u'.
The script used 'set +eu' which actually disable them!

Also set pipefail option to exit whenever a command fail in a pipeline
of commands.

Change-Id: I32bebf5f6deed75bd8aa7fd7c3f16523c1558396