Use root to create other users
I think root generally has a better chance of being allowed to create
users than a fresh anonymous client (and we need it anyways to add the
new user to groups). Specifically, when the ConfirmEdit extension is
loaded, it requires a CAPTCHA for non-admin users to create users by
default, which breaks user creation by this library.

This call previously used root, but was changed to client in change
Iff24215771 (commit aca44300cf). It’s not clear to me why this was done,
and the commit message doesn’t seem to mention it, so I’m hoping that
this isn’t actually required, but was rather left in the change more or
less accidentally.

Change-Id: I0f99241f4c861c42f639d27d544febbcfded412f