Make package more findable on npmjs.com
The library's page on npm makes it currently surprisingly tricky to get
to the actual code: https://www.npmjs.com/package/api-testing

This change should result in a "homepage" link appearing in the sidebar.

I considered also adding a repository URL, but there we do not have one
that is both human and machine readable. I think it is better to have
none than one that shows either a 404 to a human or one that fails when
used with git.

Change-Id: Ifc04e1f1f93b181b588562c1188c5c9c30c79730