Log errors when creating an account
If creating the account fails at least twice, the sleep() is sufficient
to reach the default threshold of 2 seconds in tests, which then makes
the test fail without hints as to what really happened.

So always log the error, to make debugging easier. Do not mention the
deadlock though, since this could happen for other reasons, too.

Change-Id: I73c843b6e480c24126bdc57701037acca8a0543b