Add `force` option to `makeTag` action
When we run tests in parallel, it can be the case that multiple
tests call `makeTag` simultaneously. In this case there is a race
condition in the `makeTag` logic - the check to see whether the
tag already exists can return false, a parallel request can create
the tag, and the attempt in the original process to create the
tag then fails with 'tag-create-already-exists'.

Add a `force` option to the `makeTag` action to allow "blind"
creation of tags. The API request to create the tag will always
be made regardless of whether it exists on the server, and the
request will not return error in the event
'tag-create-already-exists' is encountered. In this case, listing
the tags from the server before executing the `managetags` API
call will be skipped because we don't care about what's already
on the server.

Bug: T368903
Change-Id: I786f3c308f513ea438eda7d4bacd164a7d7db37f