Make config a module object rather than a function
Until now, require('./config') would return a function, which was called
immediately to load the configuration.  This mean that any mocks needed
to be injected before this function was called. However, initializing
the mocks outside the Mocha's describe() context caused problems with
the order of initialization, because mocks from different tests
interfered with eachother.

This patch refactors the config module to use lazy loading. That way,
mocks can be injected after other modules have already been
initialized (as long as the other modules haven't been used for anything
that would have cuased the configuration to be loaded).

Note that this changes how the config module gets imported, but since the config module is internal and not exposed by the api-testing package, that should not be a probelm.

Change-Id: I1bd95100ac4679acbd1243f6e1612a7d2776b059